% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri_value.R
\name{peri_value}
\alias{peri_value}
\title{General rounding for tables}
\usage{
peri_value(x, rspec = NULL)
}
\arguments{
\item{x}{a vector of numeric values.}

\item{rspec}{a \code{rounding_specification} object. If no \code{rspec}
is given, a default setting will round values to decimal places
based on the magnitude of the values.}
}
\value{
a vector of character values (rounded numbers).
}
\description{
\code{peri_value()} casts numeric vectors into character vectors.
The main purpose of \code{peri_value()} is to round and format
numeric data for presentation.
}
\examples{

peri_value(0.123)
peri_value(1.23)
peri_value(12.3)

with(mtcars, peri_value(disp))

}
\seealso{
Other table helpers: 
\code{\link{peri_ester}()},
\code{\link{peri_glue}()},
\code{\link{peri_pvalue}()}
}
\concept{table helpers}
