% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.glue-package.R
\docType{package}
\name{periglue-package}
\alias{periglue}
\alias{periglue-package}
\title{periglue: Make and Apply Customized Rounding Specifications for Tables}
\description{
Translate double and integer valued data into character values formatted for tabulation in manuscripts or other types of academic reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/perisphere-rwe/periglue}
  \item Report bugs at \url{https://github.com/perisphere-rwe/periglue/issues}
}

}
\author{
\strong{Maintainer}: Byron Jaeger \email{bjaeger@wakehealth.edu} (\href{https://orcid.org/0000-0001-7399-2299}{ORCID})

}
\keyword{internal}
