% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_spec.R
\name{round_spec}
\alias{round_spec}
\title{Make a rounding specification}
\usage{
round_spec(force_default = FALSE)
}
\arguments{
\item{force_default}{a logical value. If \code{TRUE}, then \code{round_spec()}
ignores any stored default (see \link{set_default_rounding_spec}) and global
options, returning factory default values. If \code{FALSE}, \code{round_spec()}
first checks for a stored default, then falls back to \verb{periglue.*}
global options.}
}
\value{
an object of class \code{rounding_specification}.
}
\description{
\code{round_spec()} creates a rounding specification object with default
settings. The settings of a rounding specification object can be
updated using functions in the \code{round_} (see \link{round_half_up},
\link{round_half_even}, \link{round_using_signif}, \link{round_using_decimal},
and \link{round_using_magnitude}) and \code{format_} (see \link{format_missing},
\link{format_big}, \link{format_small}, and \link{format_decimal}) families.
}
\details{
Rounding specifications are meant to be passed into the \link{peri_glue}
and \link{peri_value} functions. A session-wide default specification can be
stored with \code{\link[=set_default_rounding_spec]{set_default_rounding_spec()}}, which \code{round_spec()} will
use automatically when \code{force_default = FALSE}.

The \code{round_spec()} function intentionally uses no input arguments.
This is to encourage users to develop rounding specifications
using the \code{round_} and \code{format_} families in conjunction with
the pipe (\verb{|>}) operator.
}
\examples{

rspec <- round_spec()

peri_value(x = pi, rspec)

}
